+++ CudaTest/Shading.h
@@ -13,7 +13,7 @@ CUDA_HOST_DEVICE Tuple shadeHit(World* world, const HitInfo& hitInfo, int32_t re
 CUDA_HOST_DEVICE Tuple colorAt(World* world, const Ray& ray, int32_t remaining);
 CUDA_HOST_DEVICE Tuple reflectedColor(World* world, const HitInfo& inHitInfo, int32_t remaining);
 CUDA_HOST_DEVICE Tuple refractedColor(World* world, const HitInfo& inHitInfo, int32_t remaining);
-CUDA_HOST_DEVICE double schlick(const HitInfo& hitInfo);
+CUDA_HOST_DEVICE Float schlick(const HitInfo& hitInfo);
 
 CUDA_HOST_DEVICE Tuple lighting(Material* material, Shape* object, Light* light,
                const Tuple& position, const Tuple& viewDirection,
